tcmode-default: eliminate needless gdb, qemu defaults
Manually setting preferred version/provider is not required so remove
that for gdb and qemu as a first step of cleaning up this file.

This PREFERRED_VERSION list dates back to 2006, which is an era where
several versions of a given recipe was a common occurence.

(From OE-Core rev: 5f0a71d2c89f2d02e7d077c515e8d92c78824b2e)

Signed-off-by: Randy MacLeod <Randy.MacLeod@windriver.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>