oeqa/sdk/meson: Drop unreliable compiler test
The SDK Meson compiler test does not work in certain scenarios like
meta-ide-support or NATIVELSBSTRING = "debian-11":
```
Traceback (most recent call last):
   File "/srv/pokybuild/yocto-worker/qemux86/build/meta/lib/oeqa/sdk/cases/meson.py", line 47, in test_epoxy
     self.assertNotEqual(data["build"]["c"]["exelist"], data["host"]["c"]["exelist"])
                         ~~~~~~~~~~~~~^^^^^
KeyError: 'c'
```

The root cause is elusive. The contents of `meson introspect --compilers`
is no different from the working case of the standalone SDK with
NATIVELSBSTRING = "universal". Notably, the "build" field is in fact
empty, so it's almost as if the failures are correct and the success is
wrong.

The benefits of this part of the test are marginal, so fix it by
removing it.

(From OE-Core rev: 277a03dbd01a7bc9863d2ae6bbe10018b47115e8)

Signed-off-by: Tom Hochstein <tom.hochstein@oss.nxp.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>